requirements.txt
absl-py==2.1.0
astor==0.8.1
backcall==0.2.0
box2d-py==2.3.5
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
cloudpickle==2.2.1
colorama==0.4.6
debugpy==1.7.0
decorator==5.1.1
entrypoints==0.4
gast==0.2.2
google-pasta==0.2.0
grpcio==1.62.1
gym==0.26.2
gym-notices==0.0.8
h5py==2.10.0
importlib-metadata==6.7.0
ipykernel==6.16.2
ipython==7.34.0
jedi==0.19.1
jupyter_client==7.4.9
jupyter_core==4.12.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.4.4
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
numpy==1.21.6
opencv-contrib-python==4.9.0.80
opt-einsum==3.3.0
packaging==24.0
pandas==1.1.5
parso==0.8.4
pickleshare==0.7.5
prompt-toolkit==3.0.43
protobuf==3.20.1
psutil==5.9.8
pygame==2.1.0
Pygments==2.17.2
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
pyzmq==26.0.0
six==1.16.0
swig==4.2.1
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==2.3.0
tornado==6.2
traitlets==5.9.0
typing_extensions==4.7.1
wcwidth==0.2.13
Werkzeug==2.2.3
wincertstore==0.2
wrapt==1.16.0
zipp==3.15.0
